Applying fixes + updating goldens
