+++ gapic/generator/generator.py
@@ -274,9 +274,15 @@ class Generator:
                 ):
                     continue
 
+                has_requests = any(enum.name.endswith('Request') for enum in proto.enums.values())
+                has_requests |= any(message.name.endswith('Request') for message in proto.messages.values())
+                has_responses = any(enum.name.endswith('Response') for enum in proto.enums.values())
+                has_responses |= any(message.name.endswith('Response') for message in proto.messages.values())
+
                 answer.update(
                     self._get_file(
-                        template_name, api_schema=api_schema, proto=proto, opts=opts, snippet_index=snippet_index
+                        template_name, api_schema=api_schema, proto=proto, opts=opts, snippet_index=snippet_index,
+                        has_requests=has_requests, has_responses=has_responses
                     )
                 )
