+++ tests/integration/goldens/asset/google/cloud/asset_v1/types/assets/__init__.py
