+++ tests/integration/goldens/credentials/google/iam/credentials_v1/types/iamcredentials/__init__.py
