+++ tests/integration/goldens/credentials/google/iam/credentials_v1/types/iamcredentials/requests.py
@@ -1,0 +1,29 @@
+# -*- coding: utf-8 -*-
+# Copyright 2022 Google LLC
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#     http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+#
+
+
+__protobuf__ = proto.module(
+    package='google.iam.credentials.v1',
+    manifest={
+    },
+)
+
+
+__manifest__ = (
+)
+
+
+__all__ = tuple(sorted(__manifest__))
