+++ tests/integration/goldens/eventarc/google/cloud/eventarc_v1/types/trigger/__init__.py
