+++ tests/integration/goldens/logging/google/cloud/logging_v2/types/log_entry/__init__.py
