+++ tests/integration/goldens/logging/google/cloud/logging_v2/types/log_entry/requests.py
@@ -1,0 +1,37 @@
+# -*- coding: utf-8 -*-
+# Copyright 2022 Google LLC
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#     http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+#
+import proto  # type: ignore
+
+from google.api import monitored_resource_pb2  # type: ignore
+from google.logging.type import http_request_pb2  # type: ignore
+from google.logging.type import log_severity_pb2  # type: ignore
+from google.protobuf import any_pb2  # type: ignore
+from google.protobuf import struct_pb2  # type: ignore
+from google.protobuf import timestamp_pb2  # type: ignore
+
+
+__protobuf__ = proto.module(
+    package='google.logging.v2',
+    manifest={
+    },
+)
+
+
+__manifest__ = (
+)
+
+
+__all__ = tuple(sorted(__manifest__))
