+++ tests/integration/goldens/logging/google/cloud/logging_v2/types/logging/requests.py
@@ -26,17 +26,21 @@ __protobuf__ = proto.module(
     manifest={
         'DeleteLogRequest',
         'WriteLogEntriesRequest',
-        'WriteLogEntriesResponse',
-        'WriteLogEntriesPartialErrors',
         'ListLogEntriesRequest',
-        'ListLogEntriesResponse',
         'ListMonitoredResourceDescriptorsRequest',
-        'ListMonitoredResourceDescriptorsResponse',
         'ListLogsRequest',
-        'ListLogsResponse',
         'TailLogEntriesRequest',
-        'TailLogEntriesResponse',
     },
+)
+
+
+__manifest__ = (
+        'DeleteLogRequest',
+        'WriteLogEntriesRequest',
+        'ListLogEntriesRequest',
+        'ListMonitoredResourceDescriptorsRequest',
+        'ListLogsRequest',
+        'TailLogEntriesRequest',
 )
 
 
@@ -60,6 +64,7 @@ class DeleteLogRequest(proto.Message):
             For more information about log names, see
             [LogEntry][google.logging.v2.LogEntry].
     """
+    __module__ = __module__.rsplit('.', maxsplit=1)[0]  # type: ignore
 
     log_name = proto.Field(
         proto.STRING,
@@ -157,6 +162,7 @@ class WriteLogEntriesRequest(proto.Message):
             whether the logging API endpoints are working
             properly before sending valuable data.
     """
+    __module__ = __module__.rsplit('.', maxsplit=1)[0]  # type: ignore
 
     log_name = proto.Field(
         proto.STRING,
@@ -184,33 +190,6 @@ class WriteLogEntriesRequest(proto.Message):
     dry_run = proto.Field(
         proto.BOOL,
         number=6,
-    )
-
-
-class WriteLogEntriesResponse(proto.Message):
-    r"""Result returned from WriteLogEntries.
-    """
-
-
-class WriteLogEntriesPartialErrors(proto.Message):
-    r"""Error details for WriteLogEntries with partial success.
-
-    Attributes:
-        log_entry_errors (Mapping[int, google.rpc.status_pb2.Status]):
-            When ``WriteLogEntriesRequest.partial_success`` is true,
-            records the error status for entries that were not written
-            due to a permanent error, keyed by the entry's zero-based
-            index in ``WriteLogEntriesRequest.entries``.
-
-            Failed requests for which no entries are written will not
-            include per-entry errors.
-    """
-
-    log_entry_errors = proto.MapField(
-        proto.INT32,
-        proto.MESSAGE,
-        number=1,
-        message=status_pb2.Status,
     )
 
 
@@ -269,6 +248,7 @@ class ListLogEntriesRequest(proto.Message):
             the previous response. The values of other method parameters
             should be identical to those in the previous call.
     """
+    __module__ = __module__.rsplit('.', maxsplit=1)[0]  # type: ignore
 
     resource_names = proto.RepeatedField(
         proto.STRING,
@@ -292,46 +272,6 @@ class ListLogEntriesRequest(proto.Message):
     )
 
 
-class ListLogEntriesResponse(proto.Message):
-    r"""Result returned from ``ListLogEntries``.
-
-    Attributes:
-        entries (Sequence[google.cloud.logging_v2.types.LogEntry]):
-            A list of log entries. If ``entries`` is empty,
-            ``nextPageToken`` may still be returned, indicating that
-            more entries may exist. See ``nextPageToken`` for more
-            information.
-        next_page_token (str):
-            If there might be more results than those appearing in this
-            response, then ``nextPageToken`` is included. To get the
-            next set of results, call this method again using the value
-            of ``nextPageToken`` as ``pageToken``.
-
-            If a value for ``next_page_token`` appears and the
-            ``entries`` field is empty, it means that the search found
-            no log entries so far but it did not have time to search all
-            the possible log entries. Retry the method with this value
-            for ``page_token`` to continue the search. Alternatively,
-            consider speeding up the search by changing your filter to
-            specify a single log name or resource type, or to narrow the
-            time range of the search.
-    """
-
-    @property
-    def raw_page(self):
-        return self
-
-    entries = proto.RepeatedField(
-        proto.MESSAGE,
-        number=1,
-        message=log_entry.LogEntry,
-    )
-    next_page_token = proto.Field(
-        proto.STRING,
-        number=2,
-    )
-
-
 class ListMonitoredResourceDescriptorsRequest(proto.Message):
     r"""The parameters to ListMonitoredResourceDescriptors
 
@@ -348,40 +288,13 @@ class ListMonitoredResourceDescriptorsRequest(proto.Message):
             the previous response. The values of other method parameters
             should be identical to those in the previous call.
     """
+    __module__ = __module__.rsplit('.', maxsplit=1)[0]  # type: ignore
 
     page_size = proto.Field(
         proto.INT32,
         number=1,
     )
     page_token = proto.Field(
-        proto.STRING,
-        number=2,
-    )
-
-
-class ListMonitoredResourceDescriptorsResponse(proto.Message):
-    r"""Result returned from ListMonitoredResourceDescriptors.
-
-    Attributes:
-        resource_descriptors (Sequence[google.api.monitored_resource_pb2.MonitoredResourceDescriptor]):
-            A list of resource descriptors.
-        next_page_token (str):
-            If there might be more results than those appearing in this
-            response, then ``nextPageToken`` is included. To get the
-            next set of results, call this method again using the value
-            of ``nextPageToken`` as ``pageToken``.
-    """
-
-    @property
-    def raw_page(self):
-        return self
-
-    resource_descriptors = proto.RepeatedField(
-        proto.MESSAGE,
-        number=1,
-        message=monitored_resource_pb2.MonitoredResourceDescriptor,
-    )
-    next_page_token = proto.Field(
         proto.STRING,
         number=2,
     )
@@ -422,6 +335,7 @@ class ListLogsRequest(proto.Message):
             "projects/[PROJECT_ID]" "organizations/[ORGANIZATION_ID]"
             "billingAccounts/[BILLING_ACCOUNT_ID]" "folders/[FOLDER_ID]".
     """
+    __module__ = __module__.rsplit('.', maxsplit=1)[0]  # type: ignore
 
     parent = proto.Field(
         proto.STRING,
@@ -438,35 +352,6 @@ class ListLogsRequest(proto.Message):
     resource_names = proto.RepeatedField(
         proto.STRING,
         number=8,
-    )
-
-
-class ListLogsResponse(proto.Message):
-    r"""Result returned from ListLogs.
-
-    Attributes:
-        log_names (Sequence[str]):
-            A list of log names. For example,
-            ``"projects/my-project/logs/syslog"`` or
-            ``"organizations/123/logs/cloudresourcemanager.googleapis.com%2Factivity"``.
-        next_page_token (str):
-            If there might be more results than those appearing in this
-            response, then ``nextPageToken`` is included. To get the
-            next set of results, call this method again using the value
-            of ``nextPageToken`` as ``pageToken``.
-    """
-
-    @property
-    def raw_page(self):
-        return self
-
-    log_names = proto.RepeatedField(
-        proto.STRING,
-        number=3,
-    )
-    next_page_token = proto.Field(
-        proto.STRING,
-        number=2,
     )
 
 
@@ -508,6 +393,7 @@ class TailLogEntriesRequest(proto.Message):
             0-60000 milliseconds. Defaults to 2000
             milliseconds.
     """
+    __module__ = __module__.rsplit('.', maxsplit=1)[0]  # type: ignore
 
     resource_names = proto.RepeatedField(
         proto.STRING,
@@ -524,63 +410,4 @@ class TailLogEntriesRequest(proto.Message):
     )
 
 
-class TailLogEntriesResponse(proto.Message):
-    r"""Result returned from ``TailLogEntries``.
-
-    Attributes:
-        entries (Sequence[google.cloud.logging_v2.types.LogEntry]):
-            A list of log entries. Each response in the stream will
-            order entries with increasing values of
-            ``LogEntry.timestamp``. Ordering is not guaranteed between
-            separate responses.
-        suppression_info (Sequence[google.cloud.logging_v2.types.TailLogEntriesResponse.SuppressionInfo]):
-            If entries that otherwise would have been
-            included in the session were not sent back to
-            the client, counts of relevant entries omitted
-            from the session with the reason that they were
-            not included. There will be at most one of each
-            reason per response. The counts represent the
-            number of suppressed entries since the last
-            streamed response.
-    """
-
-    class SuppressionInfo(proto.Message):
-        r"""Information about entries that were omitted from the session.
-
-        Attributes:
-            reason (google.cloud.logging_v2.types.TailLogEntriesResponse.SuppressionInfo.Reason):
-                The reason that entries were omitted from the
-                session.
-            suppressed_count (int):
-                A lower bound on the count of entries omitted due to
-                ``reason``.
-        """
-        class Reason(proto.Enum):
-            r"""An indicator of why entries were omitted."""
-            REASON_UNSPECIFIED = 0
-            RATE_LIMIT = 1
-            NOT_CONSUMED = 2
-
-        reason = proto.Field(
-            proto.ENUM,
-            number=1,
-            enum='TailLogEntriesResponse.SuppressionInfo.Reason',
-        )
-        suppressed_count = proto.Field(
-            proto.INT32,
-            number=2,
-        )
-
-    entries = proto.RepeatedField(
-        proto.MESSAGE,
-        number=1,
-        message=log_entry.LogEntry,
-    )
-    suppression_info = proto.RepeatedField(
-        proto.MESSAGE,
-        number=2,
-        message=SuppressionInfo,
-    )
-
-
-__all__ = tuple(sorted(__protobuf__.manifest))
+__all__ = tuple(sorted(__manifest__))
