+++ tests/integration/goldens/redis/google/cloud/redis_v1/types/cloud_redis/responses.py
@@ -1,0 +1,79 @@
+# -*- coding: utf-8 -*-
+# Copyright 2022 Google LLC
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#     http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+#
+import proto  # type: ignore
+
+from google.protobuf import field_mask_pb2  # type: ignore
+from google.protobuf import timestamp_pb2  # type: ignore
+
+__protobuf__ = proto.module(
+    package='google.cloud.redis.v1',
+    manifest={
+        'ListInstancesResponse',
+    },
+)
+
+__manifest__ = (
+        'ListInstancesResponse',
+)
+
+
+class ListInstancesResponse(proto.Message):
+    r"""Response for
+    [ListInstances][google.cloud.redis.v1.CloudRedis.ListInstances].
+
+    Attributes:
+        instances (Sequence[google.cloud.redis_v1.types.Instance]):
+            A list of Redis instances in the project in the specified
+            location, or across all locations.
+
+            If the ``location_id`` in the parent field of the request is
+            "-", all regions available to the project are queried, and
+            the results aggregated. If in such an aggregated query a
+            location is unavailable, a dummy Redis entry is included in
+            the response with the ``name`` field set to a value of the
+            form
+            ``projects/{project_id}/locations/{location_id}/instances/``-
+            and the ``status`` field set to ERROR and ``status_message``
+            field set to "location not available for ListInstances".
+        next_page_token (str):
+            Token to retrieve the next page of results,
+            or empty if there are no more results in the
+            list.
+        unreachable (Sequence[str]):
+            Locations that could not be reached.
+    """
+    __module__ = __module__.rsplit('.', maxsplit=1)[0]  # type: ignore
+
+    @property
+    def raw_page(self):
+        return self
+
+    instances = proto.RepeatedField(
+        proto.MESSAGE,
+        number=1,
+        message='Instance',
+    )
+    next_page_token = proto.Field(
+        proto.STRING,
+        number=2,
+    )
+    unreachable = proto.RepeatedField(
+        proto.STRING,
+        number=3,
+    )
+
+
+__all__ = tuple(sorted(__manifest__))
